feat(composedmodal): move tests into one
